[Rename] : Move after creating a package
