+++ src/main/java/com/server/esgcafe/domain/dto/UserReward/UserRewardInfo.java
@@ -1,4 +1,4 @@
-package com.server.esgcafe.domain.dto;
+package com.server.esgcafe.domain.dto.UserReward;
 
 import com.server.esgcafe.domain.entity.UserReward;
 import lombok.*;
