+++ src/main/java/com/server/esgcafe/domain/dto/UserReward/UserRewardSaveRequest.java
@@ -1,4 +1,4 @@
-package com.server.esgcafe.domain.dto;
+package com.server.esgcafe.domain.dto.UserReward;
 
 import lombok.*;
